requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
pillow==10.1.0
google-genai>=1.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
aiofiles==23.2.1
aiosqlite==0.19.0
